Change scanPos to unsigned long long

+++ velox/dwio/text/reader/TextReaderImpl.cpp
@@ -498,7 +498,7 @@ T TextRowReaderImpl::getInteger(
   }
 
   int64_t v = 0;
-  uint64_t scanPos = 0;
+  unsigned long long scanPos = 0;
   errno = 0;
   auto scanCount = sscanf(s.data(), "%" SCNd64 "%lln", &v, &scanPos);
   if (scanCount != 1 || errno == ERANGE) {
@@ -615,7 +615,7 @@ float TextRowReaderImpl::getFloat(
   }
 
   float v = 0.0;
-  uint64_t scanPos = 0;
+  unsigned long long scanPos = 0;
   // We ignore ERANGE, since denormalized floats and
   // infinities are acceptable.
   auto scanCount = sscanf(str.data(), "%f%lln", &v, &scanPos);
@@ -653,7 +653,7 @@ double TextRowReaderImpl::getDouble(
     return 0.0;
   }
   double v = 0.0;
-  uint64_t scanPos = 0;
+  unsigned long long scanPos = 0;
   // We ignore ERANGE, since denormalized doubles and
   // infinities are acceptable.
   auto scanCount = sscanf(str.data(), "%lf%lln", &v, &scanPos);
